[circt] Print CIRCT errors to stderr, not stdout
When CIRCT (MFC) hits an error, print the CIRCT error to stderr instead
of stdout.  This is pragmatically done to work around an issue with mdoc
where, when it fails, it only prints whatever is on stdout.  More
basically, printing errors to stderr probably makes sense.

No test is included for this as tests of stdout/stderr are notoriously
difficult to do in a thread-safe way that is compatible with ScalaTest.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>